Add buttons for changing compute requirement.
One issue still remains. There might be a substantial delay
between the RepoChangeComputeRequirement being approved
and the boolean changing in the database.
This makes it difficult to update the UI accurately.
So, for now, we force a refresh after a 5 second delay. Ugly but works.
Will work on making it smoother...